add the capability to skip analysis of some traces (with many issues)
Added the capability to skip analysis of some traces via the traces excel. By changing the ananlyze boolan to false, the analysis of that trace will be skiped and not included in the pdf. Though, this has many issues, for instance if a whole file is skipped the program throws an error, and it is very badly implemented. With more perspective on how the script works I will come up with a better implementation later